+++ CHANGELOG.md
@@ -21,6 +21,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Fixed an issue when `bDynamicPage` is deactivated on initializing `iBlock`
 * Fixed an issue when activation events won't propagate to child components `super/i-block/modules/activation`
+* Fixed the scroll restoring after a transition `bRouter`
 
 ## v3.0.0-rc.191 (2021-05-24)