+++ CHANGELOG.md
@@ -11,6 +11,24 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :rocket: New Feature
+
+* Rewritten with a new keep-alive strategy `bDynamicPage`
+* Added a new event `asyncRenderChunkComplete` `super/i-block/modules/async-render`
+
+#### :bug: Bug Fix
+
+* Fixed an issue when `bDynamicPage` is deactivated on initializing `iBlock`
+* Fixed an issue when activation events won't propagate to child components `super/i-block/modules/activation`
+* Fixed the scroll restoring after a transition `bRouter`
+* Prevented the infinity loop when passing non-iterable objects to `iterate` `super/i-block/modules/async-render`
+
+#### :memo: Documentation
+
+* Added documentation `bDynamicPage`
+
 ## v3.0.0-rc.191 (2021-05-24)
 
 #### :rocket: New Feature